refactor: battle 정보 조회 관련 usecase 하나로 통합 #328
